Adjustments to class package structure (ended up reverting to original structure). Fixed usage of winding array in transformerobj so that the class is correctly identified.
Using objects as properties of objects from the same package is for some reason not mentioned in the ML Documentation. It was a massive waste of time searching for the solution and making attempts. Objects containing other objects is one of the first things you learn in an introduction to OOP, but somehow this is a niche use case in the ML world.