A few minor changes to clean things up. Had briefly experimented with outputting state activity, but it's useless under Super Step semantics, which are important to the operation of the model.
The solution going forward is to read state activity through broadcasting/logging. Unfortunately, this doesn't come with a useful time stamp (because of Super Step), so it will be accompanied with an awareness of possible transitions from each state (aka a manually-built copy of this information).

Before this works in a useful capacity, there needs to be division between control mode logic and decision-related logic (a revision of the state breakdown and labeling).